feat: Adding the dart test command
