+++ lib/src/cli/templates/templates.dart
