+++ lib/src/cli/templates/test_optimizer_bundle.dart
