+++ lib/src/commands/commands.dart
@@ -1,5 +1,6 @@
 export 'create/commands/commands.dart';
 export 'create/create.dart';
+export 'dart/dart.dart';
 export 'packages/packages.dart';
 export 'test/test.dart';
 export 'update.dart';
