+++ lib/src/commands/dart/commands/dart_test_commands.dart
@@ -1,3 +1,4 @@
+import 'dart:io';
 import 'dart:math';
 
 import 'package:args/args.dart';
@@ -5,15 +6,13 @@ import 'package:args/command_runner.dart';
 import 'package:mason/mason.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
-import 'package:universal_io/io.dart';
 import 'package:very_good_cli/src/cli/cli.dart';
 
-/// Signature for the [Flutter.installed] method.
-typedef FlutterInstalledCommand =
-    Future<bool> Function({required Logger logger});
+/// Signature for the [Dart.installed] method.
+typedef DartInstalledCommand = Future<bool> Function({required Logger logger});
 
-/// Signature for the [Flutter.test] method.
-typedef FlutterTestCommand =
+/// Signature for the [Dart.test] method.
+typedef DartTestCommandCall =
     Future<List<int>> Function({
       required Logger logger,
       String cwd,
@@ -29,18 +28,18 @@ typedef FlutterTestCommand =
       void Function(String)? stderr,
     });
 
-/// {@template test_command}
-/// `very_good test` command for running tests.
+/// {@template dart_test_command}
+/// `very_good dart test` command for running dart tests.
 /// {@endtemplate}
-class TestCommand extends Command<int> {
-  /// {@macro test_command}
-  TestCommand({
-    required Logger logger,
-    @visibleForTesting FlutterInstalledCommand? flutterInstalled,
-    @visibleForTesting FlutterTestCommand? flutterTest,
-  }) : _logger = logger,
-       _flutterInstalled = flutterInstalled ?? Flutter.installed,
-       _flutterTest = flutterTest ?? Flutter.test {
+class DartTestCommand extends Command<int> {
+  /// {@macro packages_command}
+  DartTestCommand({
+    Logger? logger,
+    DartTestCommandCall? dartTest,
+    DartInstalledCommand? dartInstalled,
+  }) : _logger = logger ?? Logger(),
+       _dartTest = dartTest ?? Dart.test,
+       _dartInstalled = dartInstalled ?? Dart.installed {
     argParser
       ..addFlag(
         'coverage',
@@ -91,51 +90,21 @@ class TestCommand extends Command<int> {
             'within test files.',
       )
       ..addFlag(
-        'update-goldens',
-        help:
-            'Whether "matchesGoldenFile()" calls within your test methods '
-            'should update the golden files.',
-        negatable: false,
-      )
-      ..addFlag(
         'force-ansi',
         defaultsTo: null,
         help:
             'Whether to force ansi output. If not specified, '
             'it will maintain the default behavior based on stdout and stderr.',
         negatable: false,
-      )
-      ..addMultiOption(
-        'dart-define',
-        help:
-            'Additional key-value pairs that will be available as constants '
-            'from the String.fromEnvironment, bool.fromEnvironment, '
-            'int.fromEnvironment, and double.fromEnvironment constructors. '
-            'Multiple defines can be passed by repeating '
-            '"--dart-define" multiple times.',
-        valueHelp: 'foo=bar',
-      )
-      ..addMultiOption(
-        'dart-define-from-file',
-        help:
-            'The path of a .json or .env file containing key-value pairs '
-            'that will be available as environment variables. '
-            'These can be accessed using the String.fromEnvironment, '
-            'bool.fromEnvironment, and int.fromEnvironment constructors. '
-            'Multiple defines can be passed by repeating '
-            '"--dart-define-from-file" multiple times. '
-            'Entries from "--dart-define" with identical keys take '
-            'precedence over entries from these files.',
-        valueHelp: 'use-define-config.json|.env',
       );
   }
 
   final Logger _logger;
-  final FlutterInstalledCommand _flutterInstalled;
-  final FlutterTestCommand _flutterTest;
+  final DartTestCommandCall _dartTest;
+  final DartInstalledCommand _dartInstalled;
 
   @override
-  String get description => 'Run tests in a Dart or Flutter project.';
+  String get description => 'Command for managing packages.';
 
   @override
   String get name => 'test';
@@ -166,7 +135,7 @@ This command should be run from the root of your Flutter project.''');
     );
     final excludeTags = _argResults['exclude-tags'] as String?;
     final tags = _argResults['tags'] as String?;
-    final isFlutterInstalled = await _flutterInstalled(logger: _logger);
+    final isDartInstalled = await _dartInstalled(logger: _logger);
     final excludeFromCoverage = _argResults['exclude-coverage'] as String?;
     final randomOrderingSeed =
         _argResults['test-randomize-ordering-seed'] as String?;
@@ -174,20 +143,14 @@ This command should be run from the root of your Flutter project.''');
         ? Random().nextInt(4294967295).toString()
         : randomOrderingSeed;
     final optimizePerformance = _argResults['optimization'] as bool;
-    final updateGoldens = _argResults['update-goldens'] as bool;
     final forceAnsi = _argResults['force-ansi'] as bool?;
-    final dartDefine = _argResults['dart-define'] as List<String>?;
-    final dartDefineFromFile =
-        _argResults['dart-define-from-file'] as List<String>?;
     final rest = _argResults.rest;
 
-    if (isFlutterInstalled) {
+    if (isDartInstalled) {
       try {
-        final results = await _flutterTest(
+        final results = await _dartTest(
           optimizePerformance:
-              optimizePerformance &&
-              !TestCLIRunner.isTargettingTestFiles(rest) &&
-              !updateGoldens,
+              optimizePerformance && !TestCLIRunner.isTargettingTestFiles(rest),
           recursive: recursive,
           logger: _logger,
           stdout: _logger.write,
@@ -200,14 +163,7 @@ This command should be run from the root of your Flutter project.''');
           arguments: [
             if (excludeTags != null) ...['-x', excludeTags],
             if (tags != null) ...['-t', tags],
-            if (updateGoldens) '--update-goldens',
-            if (dartDefine != null)
-              for (final value in dartDefine) '--dart-define=$value',
-            if (dartDefineFromFile != null)
-              for (final value in dartDefineFromFile)
-                '--dart-define-from-file=$value',
             ...['-j', concurrency],
-            '--no-pub',
             ...rest,
           ],
         );
